+++ lib/src/flutter/word.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/gestures.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 
 typedef WordWidgetCallback = void Function(String text, String id);
@@ -6,12 +8,14 @@ class WordWidget extends LeafRenderObjectWidget {
   final String text;
   final int id;
   final TextStyle style;
+  final VoidCallback? onTap;
 
   const WordWidget({
     super.key,
     required this.text,
     required this.id,
     this.style = const TextStyle(color: Color(0xFF000000), fontSize: 14),
+    this.onTap,
   });
 
   @override
@@ -21,6 +25,7 @@ class WordWidget extends LeafRenderObjectWidget {
       id: id,
       style: style,
       textDirection: Directionality.of(context),
+      onTap: onTap,
     );
   }
 
@@ -33,7 +38,8 @@ class WordWidget extends LeafRenderObjectWidget {
       ..text = text
       ..id = id
       ..style = style
-      ..textDirection = Directionality.of(context);
+      ..textDirection = Directionality.of(context)
+      ..onTap = onTap;
   }
 }
 
@@ -43,17 +49,31 @@ class RenderWord extends RenderBox {
     required int id,
     required TextStyle style,
     required TextDirection textDirection,
+    VoidCallback? onTap,
   }) : _text = text,
        _id = id,
        _style = style,
-       _textDirection = textDirection {
+       _textDirection = textDirection,
+       _onTap = onTap {
     _textPainter = TextPainter(
       text: TextSpan(text: _text, style: _style),
       textDirection: _textDirection,
     );
+    _tapRecognizer = TapGestureRecognizer()
+      ..onTap = () {
+        _onTap?.call();
+      };
   }
 
   late final TextPainter _textPainter;
+  late final TapGestureRecognizer _tapRecognizer;
+
+  VoidCallback? _onTap;
+  VoidCallback? get onTap => _onTap;
+  set onTap(VoidCallback? value) {
+    if (_onTap == value) return;
+    _onTap = value;
+  }
 
   String _text;
   String get text => _text;
@@ -104,5 +124,18 @@ class RenderWord extends RenderBox {
   @override
   bool hitTestSelf(Offset position) {
     return true;
+  }
+
+  @override
+  void handleEvent(PointerEvent event, covariant BoxHitTestEntry entry) {
+    if (event is PointerDownEvent && _onTap != null) {
+      _tapRecognizer.addPointer(event);
+    }
+  }
+
+  @override
+  void dispose() {
+    _tapRecognizer.dispose();
+    super.dispose();
   }
 }